refactor db insertion of intg test
