+++ master/test/olddata/pre_remove_steps.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 
 	"github.com/determined-ai/determined/master/internal/db"
+	"github.com/determined-ai/determined/master/test/testutils"
 )
 
 //go:embed pre_remove_steps.sql
@@ -19,7 +20,7 @@ const migration001213 = 20200630141158
 
 // PreRemoveStepsExperimentsData holds the migration and useful constants for pre_remove_steps.sql.
 type PreRemoveStepsExperimentsData struct {
-	MustMigrate                  MustMigrateFn
+	MustMigrate                  testutils.MustMigrateFn
 	CompletedSingleExpID         int32
 	CompletedAdaptiveSimpleExpID int32
 	CompletedPBTExpID            int32
@@ -30,11 +31,11 @@ type PreRemoveStepsExperimentsData struct {
 // PreRemoveStepsExperiments returns a PreRemoveStepsExperimentsData.
 func PreRemoveStepsExperiments() PreRemoveStepsExperimentsData {
 	mustMigrate := func(t *testing.T, pgdb *db.PgDB, migrationsPath string) {
-		extra := migrationExtra{
+		extra := testutils.MigrationExtra{
 			When: migration001213,
 			SQL:  preRemoveStepsSQL,
 		}
-		mustMigrateWithExtras(t, pgdb, migrationsPath, extra)
+		testutils.MustMigrateWithExtras(t, pgdb, migrationsPath, extra)
 	}
 	return PreRemoveStepsExperimentsData{
 		MustMigrate:                  mustMigrate,
